test(core): migrate IpcConnectionTests (Wave D)
4 tests, 9 failure sites preserved 1:1. Two `guard case .line(...)` destructures convert to Issue.record + return.
